Copilotservice/DRIServices/Functions/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-identity==1.13.0
Flask==2.2.5
langchain==0.0.187
openai==0.26.4
azure-search-documents==11.4.0b3
azure-storage-blob==12.14.1
